add files for taask 5

+++ 0x14-bit_manipulation/5-flip_bits.c
@@ -10,7 +10,7 @@
 
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
-	unsigned int d;
+	unsigned long int d;
 	int count = 0;
 
 	d = n ^ m;
